# docs/api/openapi_v2.json
{
  "components": {
    "schemas": {
...
      "ContextDigestBody": {
        "description": "Request body for :class:`backend.routing.contract.ContextDigest`.",
        "properties": {
          "repo": {
            "default": "",
            "title": "Repo",
            "type": "string"
          },
          "signals": {
            "$ref": "#/components/schemas/ContextSignalsBody"
          }
        },
        "title": "ContextDigestBody",
        "type": "object"
      },
      "ContextSignalsBody": {
        "description": "Request body for :class:`backend.routing.contract.ContextSignals`.",
        "properties": {
          "has_tests": {
            "default": false,
            "title": "Has Tests",
            "type": "boolean"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "title": "Languages",
            "type": "array"
          }
        },
        "title": "ContextSignalsBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrincipalResponseV2": {
        "description": "The authenticated caller, as returned by ``GET /v2/auth/me``.",
        "properties": {
          "authMethod": {
            "title": "Authmethod",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "title": "Roles",
            "type": "array"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "tenantId": {
            "title": "Tenantid",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "tenantId",
          "roles",
          "scopes",
          "authMethod"
        ],
        "title": "PrincipalResponseV2",
        "type": "object"
      },
...
      "Role": {
        "description": "Canonical, emitted Control Plane roles (ADR-018 \u00a716.1.1).\n\n``author`` (the older \u00a714.2 spelling) is never a member of this enum: it\nis accepted only as a legacy input alias for :attr:`MAINTAINER` by\n:func:`backend.auth.roles.normalize_role`.",
        "enum": [
          "owner",
          "admin",
          "maintainer",
          "operator",
          "viewer"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "RouteInputBody": {
        "description": "Request body for :class:`backend.routing.contract.RouteInput`.",
        "properties": {
          "attachments": {
            "items": {
              "type": "string"
            },
            "title": "Attachments",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "RouteInputBody",
        "type": "object"
      },
      "RouteRequestBody": {
        "description": "Request body for ``POST /v2/route`` (reference \u00a79.2 ``RouteRequest``).",
        "properties": {
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "context_digest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextDigestBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "input": {
            "$ref": "#/components/schemas/RouteInputBody"
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "schemaVersion": {
            "default": "1.0",
            "title": "Schemaversion",
            "type": "string"
          },
          "session_id": {
            "default": "",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "input"
        ],
        "title": "RouteRequestBody",
        "type": "object"
      },
...
      "ServiceCredentialCreateRequestV2": {
        "description": "Request body for ``POST /v2/auth/service-credentials``.",
        "properties": {
          "expiresInDays": {
            "maximum": 90.0,
            "minimum": 1.0,
            "title": "Expiresindays",
            "type": "integer"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Roles",
            "type": "array"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "subject": {
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "roles",
          "expiresInDays"
        ],
        "title": "ServiceCredentialCreateRequestV2",
        "type": "object"
      },
      "ServiceCredentialResponseV2": {
        "description": "A service credential, as returned by list/create/revoke.",
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "expiresAt": {
            "title": "Expiresat",
            "type": "string"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "keyId": {
            "title": "Keyid",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "title": "Roles",
            "type": "array"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "keyId",
          "subject",
          "roles",
          "scopes",
          "createdAt",
          "expiresAt"
        ],
        "title": "ServiceCredentialResponseV2",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "oidcBearer": {
        "bearerFormat": "JWT",
        "description": "OIDC-issued bearer JWT, validated against the configured JWKS.",
        "scheme": "bearer",
        "type": "http"
      },
      "serviceBearer": {
        "bearerFormat": "adk_live_<key-id>_<secret>",
        "description": "Governed service credential minted via `autodev auth service-key create`.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Browser session established by the OIDC PKCE login flow.",
        "in": "cookie",
        "name": "autodev_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "AutoDev Orchestrator - Control Plane API /v2",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/auth/me": {
      "get": {
        "description": "Return the calling principal's identity, roles, and effective scopes.\n\nArgs:\n    principal: The authenticated caller.\n\nReturns:\n    The authenticated principal's public representation.",
        "operationId": "get_me_v2_v2_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalResponseV2"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get Me V2",
        "tags": [
          "auth"
        ],
        "x-autodev-required-scope": "auth:self"
      }
    },
    "/v2/auth/oidc/callback": {
      "get": {
        "description": "Complete an OIDC login: exchange the code, mint a session.\n\nArgs:\n    code: Authorization code returned by the provider.\n    state: Anti-CSRF state matching a pending login started by\n        :func:`oidc_login_v2`.\n\nReturns:\n    A redirect to the original ``returnTo`` path with an HttpOnly,\n    Secure session cookie set.\n\nRaises:\n    HTTPException: 400 if ``state`` does not match a pending login; 502\n        if the provider rejects the code exchange.",
        "operationId": "oidc_callback_v2_v2_auth_oidc_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oidc Callback V2",
        "tags": [
          "auth"
        ],
        "x-autodev-public": true
      }
    },
    "/v2/auth/oidc/login": {
      "get": {
        "description": "Start an OIDC authorization-code + PKCE login.\n\nArgs:\n    returnTo: Relative path to redirect to after a successful login.\n        Absolute/external URLs are rejected in favor of ``/``.\n\nReturns:\n    A redirect to the OIDC provider's authorization endpoint.\n\nRaises:\n    HTTPException: 404 if OIDC is not configured.",
        "operationId": "oidc_login_v2_v2_auth_oidc_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "returnTo",
            "required": false,
            "schema": {
              "default": "/",
              "title": "Returnto",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oidc Login V2",
        "tags": [
          "auth"
        ],
        "x-autodev-public": true
      }
    },
    "/v2/auth/service-credentials": {
      "get": {
        "description": "List every service credential for the caller's tenant.\n\nArgs:\n    principal: The authenticated caller.\n\nReturns:\n    The tenant's service credentials. Never includes a secret.",
        "operationId": "list_service_credentials_v2_v2_auth_service_credentials_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceCredentialResponseV2"
                  },
                  "title": "Response List Service Credentials V2 V2 Auth Service Credentials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "List Service Credentials V2",
        "tags": [
          "auth"
        ],
        "x-autodev-required-scope": "service_credential:admin"
      },
      "post": {
        "description": "Mint a new service credential for the caller's tenant.\n\nArgs:\n    body: The credential's subject, roles, scopes, and expiry.\n    principal: The authenticated caller.\n\nReturns:\n    The persisted credential, including the one-time presented key.\n\nRaises:\n    HTTPException: 400 if the requested scopes exceed the requested\n        roles' grants.",
        "operationId": "create_service_credential_v2_v2_auth_service_credentials_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCredentialCreateRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceCredentialResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create Service Credential V2",
        "tags": [
          "auth"
        ],
        "x-autodev-required-scope": "service_credential:admin"
      }
    },
    "/v2/auth/service-credentials/{key_id}": {
      "delete": {
        "description": "Immediately revoke a service credential owned by the caller's tenant.\n\nArgs:\n    key_id: The credential's non-secret identifier.\n    principal: The authenticated caller.\n\nReturns:\n    An empty ``204`` response.\n\nRaises:\n    HTTPException: 404 if the credential does not exist or belongs to\n        another tenant.",
        "operationId": "revoke_service_credential_v2_v2_auth_service_credentials__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Revoke Service Credential V2",
        "tags": [
          "auth"
        ],
        "x-autodev-required-scope": "service_credential:admin"
      }
    },
    "/v2/auth/session": {
      "delete": {
        "description": "Log out: revoke the caller's session and clear its cookie.\n\nArgs:\n    response: Outgoing response, updated to clear the session cookie.\n    principal: The authenticated caller.\n\nReturns:\n    An empty ``204`` response.",
        "operationId": "logout_v2_v2_auth_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Logout V2",
        "tags": [
          "auth"
        ],
        "x-autodev-required-scope": "auth:self"
      }
    },
    "/v2/auth/session/refresh": {
      "post": {
        "description": "Rotate the caller's browser session using its refresh token.\n\nArgs:\n    response: Outgoing response, updated with the rotated session cookie.\n    principal: The authenticated caller.\n\nReturns:\n    The refreshed principal.\n\nRaises:\n    HTTPException: 400 if the caller does not have an active session.",
        "operationId": "refresh_session_v2_v2_auth_session_refresh_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalResponseV2"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Refresh Session V2",
        "tags": [
          "auth"
        ],
        "x-autodev-required-scope": "auth:self"
      }
    },
...
    "/v2/route": {
      "post": {
        "description": "Classify a task and return its routed decision.\n\nArgs:\n    body: A ``RouteRequest`` document (reference \u00a79.2). ``context`` is an\n        additive, out-of-contract field carrying caller-supplied signals\n        (e.g. an upstream ``intent`` hint) forwarded to the Router.\n    service: Routing service dependency.\n\nReturns:\n    A ``RouteDecision`` document as JSON.",
        "operationId": "route_request_v2_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route Request V2 Route Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "oidcBearer": []
          },
          {
            "serviceBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Route Request",
        "tags": [
          "routing"
        ],
        "x-autodev-required-scope": "flow:read"
      }
    },
...
  }
}